Fix compile errors

+++ assignment4/temp.h
@@ -20,6 +20,6 @@
 void temp_init(); /* initializes the temperature device */
 int temp_is_data_ready(); /* checks if the most recent conversion is complete */
 void temp_start(); /* starts conversion */
-signed int temp_get(); /* reads value from most recent conversion */
+signed long temp_get(); /* reads value from most recent conversion */
 
 #endif
